requirements.txt
# Dependencias para análisis y export
numpy>=1.26
pandas>=2.2
scikit-learn>=1.5
pyyaml>=6.0.1
geopandas>=1.0
shapely>=2.0
matplotlib>=3.9
requests>=2.31
